LayoutTests/http/tests/security/frameNavigation/xss-DENIED-plugin-navigation-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/resources/cross-frame-iframe.html' from frame with URL 'http://localhost:8000/security/frameNavigation/resources/frame-with-plugin-to-navigate.html'. The frame attempting navigation is neither same-origin with the target, nor is it the target's parent or opener.



--------
Frame: '<!--frame1-->'
--------

Frame-with-plugin-to-navigate

localhost

Test PASSED


--------
Frame: 'toNavigate'
--------
Inner iframe.
